Rename blockchain_data_4_2.py to raw-network-data/blockchain_data_4_2.py
**Moved extract script to new directory**
The files contained in this directory represent the raw data extracted from the Bitcoin blockchain for the respective seed address. There are two network files produced for each seed address, the cash-in and cash-out network files in JSON. The files suffixed with 'lower' provides additional data and are currently not used in any analysis for this project, however could be useful for future research.